Refactor authentication configuration and controllers
This commit refactors the authentication configuration and controllers. It replaces the usage of the cloudinary package with the passport package for handling authentication. The GoogleStrategy from passport-google-oauth20 is now used for Google authentication. The handleGoogleLogin function in the auth.controllers.js file is updated to handle the Google login process. The cloudinary configuration file is removed as it is no longer needed. The changes also include adding the auth.routes.js file to handle the Google authentication routes.